tests/test_main.rs
#[cfg(test)]
mod tests {
    use std::env;
    use std::fs::File;
    use std::io::{BufRead, BufReader, Write};
    use std::path::Path;
    use std::fs;

    fn cleanup(file_path: &str) {
        let _ = fs::remove_file(file_path);  // Ignore errors if file doesn't exist
    }

    #[test]
    fn test_main_with_file() {
        let file_path = "src/test_file.txt";
        
        // Cleanup before test
        cleanup(file_path);
        
        // Create a test file
        let mut file = File::create(file_path).unwrap();
        file.write_all(b"Hello\nWorld\n").unwrap();
        
        // Instead of directly calling main, you should either:
        // Option 1: Make main testable by accepting arguments
        super::run_with_file(file_path).unwrap();  // Assuming you create this function
        
        // Check file contents
        let reader = BufReader::new(File::open(file_path).unwrap());
        let lines: Vec<String> = reader.lines().map(|l| l.unwrap()).collect();
        assert_eq!(lines, vec!["Hello", "World"]);
        
        // Cleanup after test
        cleanup(file_path);
    }

    #[test]
    fn test_main_with_empty_file() {
        let file_path = "src/empty_file.txt";
        
        // Cleanup before test
        cleanup(file_path);
        
        // Create an empty test file
        File::create(file_path).unwrap();
        
        // Use a Result to handle the empty file case
        let result = super::run_with_file(file_path);
        assert!(result.is_err());  // Or whatever behavior you expect
        
        // Cleanup after test
        cleanup(file_path);
    }

    #[test]
    fn test_main_without_file() {
        // Test the error case when no file is provided
        let result = super::run_with_file("");
        assert!(result.is_err());
    }
}